feat(ui5-panel): implement headerAriaLabelledBy property
