+++ packages/main/src/Panel.hbs
@@ -39,11 +39,15 @@
 				role="heading"
 				aria-level="{{headerAriaLevel}}"
 				class="ui5-panel-header-title"
+				aria-labelledby="{{_id}}-headerText"
 			>
 				{{headerText}}
 			</div>
 		{{/if}}
 	</div>
+
+	<div id="aria-labelledby-texts"></div>
+	<span id="{{_id}}-headerText" class="ui5-hidden-text">{{headerAriaLabelledByText}}</span>
 
 	<!-- content area -->
 	<div class="ui5-panel-content" id="{{_id}}-content" tabindex="-1" style="{{styles.content}}">
